feat: add register callbacks command
